ads добавил

+++ uwb-firmware/Core/Inc/ps.h
@@ -3,7 +3,7 @@
 #include <stdint.h>
 
 typedef struct {
-    uint32_t pressure;
+    float pressure;
 } ps_context_t;
 
 void ps_read(ps_context_t* ps);
